+++ CHANGES.rst
@@ -1,6 +1,15 @@
-0.1.0
+[Unreleased](https://github.com/crim-ca/notebook-utils/tree/main) (latest)
+------------------------------------------------------------------------------------------------------------------------
+
+[//]: <> (Remove when new items added)
+- No changes recorded yet.
+
+[0.2.0](https://github.com/crim-ca/notebook-utils/tree/0.2.0) (11-05-2022)
+------------------------------------------------------------------------------------
+- Add function to convert a jupyter notebook to html.
+- Add functions to strip a notebook or a folder of notebooks.
+
+[0.1.0](https://github.com/crim-ca/notebook-utils/tree/0.1.0) (30-09-2021)
 ------------------------------------------------------------------------------------
 
-Features / Changes
-~~~~~~~~~~~~~~~~~~~~~
-* Add functions to verify all notebooks within a folder are stripped.
+- Add functions to verify all notebooks within a folder are stripped.